requirements.txt
requests
beautifulsoup4
dbus-python
pyecharts
rich
